Added caching of tasks
