+++ app/build.gradle
@@ -25,7 +25,7 @@ android {
 
     defaultConfig {
         applicationId "org.grassroot.android"
-        minSdkVersion 10
+        minSdkVersion 14
         targetSdkVersion 23
         versionCode 4
         versionName "0.1.4"
